requirements.txt
azure-ai-documentintelligence==1.0.0
anthropic==0.40.0
Pillow==11.0.0
python-dotenv==1.0.1
lxml==5.3.0
requests==2.32.3
easyocr==1.7.2
